+++ src/main/java/com/bitbank/controllers/v1/BankAccountTransactionsController.java
@@ -20,6 +20,7 @@ import com.bitbank.services.v1.BankAccountTransactionsService;
 @RequestMapping("/api/v1/bank-account-transactions")
 public class BankAccountTransactionsController {
 	
+	@Autowired
 	private BankAccountTransactionsService bankAccountTransactionsService;
 	
 	@Autowired