+++ src/test/java/com/bitbank/test/config/AuthTokenFilterTest.java
@@ -1,18 +1,14 @@
-package com.bitbank.test.controllers.v1;
+package com.bitbank.test.config;
 
-import static org.hamcrest.CoreMatchers.is;
 import static org.mockito.Mockito.when;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
-import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.test.context.TestPropertySource;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
@@ -25,17 +21,13 @@ import com.bitbank.utils.JwtUtils;
 
 @ExtendWith(SpringExtension.class)
 @WebMvcTest(PingController.class)
-@TestPropertySource("classpath:application.properties")
-class PingControllerTest {
+class AuthTokenFilterTest {
 
 	@Autowired
 	private MockMvc mvc;
 
 	@Autowired
 	AuthTokenFilter authTokenFilter;
-
-	@MockBean
-	AuthEntryPointJwt authEntryPointJwt;
 
 	@MockBean
 	JwtUtils jwtUtils;
@@ -45,16 +37,34 @@ class PingControllerTest {
 
 	@MockBean
 	private PingService pingService;
-
-	@Value("${app.version}")
-	private String appVersion;
+	
+	@MockBean
+	AuthEntryPointJwt authEntryPointJwt;
 
 	@Test
-	void indexShouldReturnApiVersion() throws Exception {
+	void testCanReturnNullIfJwtIsMissing() throws Exception {
+		mvc.perform(get("/api/v1/ping")).andExpect(status().isOk());
+	}
 
-		when(pingService.show()).thenReturn(appVersion);
+	@Test
+	void testCanValidateToken() throws Exception {
+		String token = "a1.b2.c3";
+		when(jwtUtils.validateJwtToken(token)).thenReturn(true);
+		mvc.perform(get("/api/v1/ping").header("Authorization", "Bearer " + token)).andExpect(status().isOk());
+	}
 
-		mvc.perform(get("/api/v1/ping")).andExpect(status().isOk()).andExpect(jsonPath("$.version", is(appVersion)));
+	@Test
+	void testCannotValidateToken() throws Exception {
+		String token = "a1.b2.c3";
+		when(jwtUtils.validateJwtToken(token)).thenReturn(false);
+		mvc.perform(get("/api/v1/ping").header("Authorization", "Bearer " + token)).andExpect(status().isOk());
+	}
+
+	@Test
+	void testCanReturnNullIfJwtIsMissingButOtherHeaderIsInPlace() throws Exception {
+		String token = "a1.b2.c3";
+		mvc.perform(get("/api/v1/ping").header("Authorization", "NotStartWithBearer " + token))
+				.andExpect(status().isOk());
 	}
 
 }